fix: release build fixes (#35)
* fix: release build fixes

I copied the release build from minikube when doing the bootc extension.
We found a few small issues and made changes, so I am contributing them
back. I've included all 4 changes since they are all small and on the
same file, but I can split them up or remove some if requested.

line 51/78: Set the git email to be the person that kicked off the
release (remove fbenoit hardcoding). Using a bot would be better, but
this works for now.

line 87: Fixes the missing version in the description of the automated
PR to bump version from e.g. "has been released." to "0.2.0 has been
released."

line 96: Change the bumped version back before doing the build so that
the correct version is in the released extension. e.g. today the
released extension would be correctly tagged 0.2.0 but once you
install it the version from package.json will say 0.3.0 (the bumped
version).

lines 131-135: Publish the image as 'latest', so that the release
build is also the latest.

Fixes #32.

Signed-off-by: Tim deBoer <git@tdeboer.ca>

* chore: remove latest from the CI job and use nightly instead

Change-Id: I60e3814d4966dade5112e4e69267a1accb3dc213
Signed-off-by: Florent Benoit <fbenoit@redhat.com>

* chore: fix date

Change-Id: Ifb23b8e4375fd82cdcb1370bf32522fe4e2b0add
Signed-off-by: Florent Benoit <fbenoit@redhat.com>

---------

Signed-off-by: Tim deBoer <git@tdeboer.ca>
Signed-off-by: Florent Benoit <fbenoit@redhat.com>
Co-authored-by: Florent Benoit <fbenoit@redhat.com>